+++ src/main/java/store_cyber_point/cyber_point/dto/ProductDto.java
@@ -3,7 +3,7 @@ package store_cyber_point.cyber_point.dto;
 import java.math.BigDecimal;
 import store_cyber_point.cyber_point.Entity.Product;
 
-public record ProductDto(Long id, String name, BigDecimal price, String description, String image) {
+public record ProductDto(Long id, String name, Double price, String description, String image) {
 
   public static ProductDto fromEntity(Product product) {
     return new ProductDto(
